Added down scripts for deleting keys for a user
Added a script for list a single user key.
Added a script to remove the deploy user key.
Clarified the README to include instructions for installing jq which
is used to parse the output of AWS command
Altered the README to include only the aws_up and aws_down commands